PD-282 Read load balancer settings from a single table
